# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Aurigraph V11 Platform API
  description: |
    High-performance blockchain platform APIs for 2M+ TPS
    
    ## Overview
    Aurigraph V11 is a next-generation blockchain platform built with Java 21, Quarkus, and GraalVM.
    It provides high-throughput transaction processing, quantum-resistant cryptography, cross-chain bridging,
    and AI-driven optimization capabilities.
    
    ## Features
    - **Ultra-High Throughput**: 2M+ transactions per second
    - **Quantum-Resistant Security**: Post-quantum cryptography implementation
    - **Cross-Chain Bridge**: Interoperability with Ethereum, Solana, and other chains
    - **AI Optimization**: Machine learning-driven consensus and performance optimization
    - **HMS Integration**: Healthcare Management System integration
    - **Real-time Monitoring**: Comprehensive performance and health monitoring
    
    ## Performance Targets
    - **TPS**: 2,000,000+ transactions per second
    - **Latency**: Sub-millisecond transaction processing
    - **Startup**: <1 second native binary startup
    - **Memory**: <256MB native binary footprint
    
    ## Technology Stack
    - **Runtime**: Java 21 with Virtual Threads
    - **Framework**: Quarkus 3.26.2
    - **Native**: GraalVM native compilation
    - **Transport**: HTTP/2 with TLS 1.3, gRPC
    - **Consensus**: HyperRAFT++ algorithm
    - **Cryptography**: CRYSTALS-Kyber/Dilithium post-quantum algorithms
    
  version: 11.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/v11/transactions:
    post:
      tags:
        - Transactions
      summary: Process a transaction
      description: Submit a transaction for processing
      operationId: processTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
            examples:
              simple_transfer:
                summary: Simple transfer transaction
                value:
                  transactionId: "tx_001_2025"
                  amount: 1000.50
      responses:
        '201':
          description: Transaction processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Invalid transaction data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service temporarily unavailable
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
...
    TransactionRequest:
      type: object
      properties:
        transactionId:
          type: string
          description: Unique transaction identifier
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 256
        amount:
          type: number
          format: double
          description: Transaction amount
          minimum: 0
          maximum: 999999999999.99
      required:
        - transactionId
        - amount
      example:
        transactionId: "tx_2025_001"
        amount: 1000.50
    TransactionResponse:
      type: object
      properties:
        transactionId:
          type: string
          description: Transaction identifier
        status:
          type: string
          enum: [PROCESSED, PENDING, FAILED]
          description: Transaction processing status
        amount:
          type: number
          format: double
          description: Transaction amount
        timestamp:
          type: integer
          format: int64
          description: Processing timestamp
        message:
          type: string
          description: Status message
      required:
        - transactionId
        - status
        - amount
        - timestamp
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        timestamp:
          type: integer
          format: int64
        details:
          type: object
          additionalProperties: true
      required:
        - error
        - timestamp
...